+++ modules/apps/my-subscriptions/my-subscriptions-web/src/main/java/com/liferay/my/subscriptions/web/internal/util/MySubscriptionsUtil.java
@@ -20,6 +20,7 @@ import com.liferay.asset.kernel.model.AssetRendererFactory;
 import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.document.library.kernel.model.DLFileEntryType;
 import com.liferay.document.library.kernel.model.DLFolder;
+import com.liferay.document.library.kernel.model.DLFolderConstants;
 import com.liferay.document.library.kernel.service.DLFileEntryTypeLocalServiceUtil;
 import com.liferay.journal.model.JournalFolder;
 import com.liferay.message.boards.kernel.model.MBCategory;
@@ -156,6 +157,13 @@ public class MySubscriptionsUtil {
 
 			return dlFileEntryType.getName(locale);
 		}
+		else if (className.equals(DLFolderConstants.getClassName()) ||
+				 className.equals(Folder.class.getName())) {
+
+			if (group != null) {
+				return LanguageUtil.get(locale, "home");
+			}
+		}
 		else if (className.equals(JournalFolder.class.getName())) {
 			if (group != null) {
 				return LanguageUtil.get(locale, "home");
@@ -198,11 +206,6 @@ public class MySubscriptionsUtil {
 			WikiNode wikiNode = WikiNodeLocalServiceUtil.getWikiNode(classPK);
 
 			return wikiNode.getName();
-		}
-		else if (className.equals(Folder.class.getName())) {
-			if (group != null) {
-				return LanguageUtil.get(locale, "home");
-			}
 		}
 
 		if (group != null) {