+++ modules/apps/web-experience/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/exportimport/portlet/preferences/processor/AssetPublisherExportImportPortletPreferencesProcessor.java
@@ -88,6 +88,13 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
+ * Provides the implementation of
+ * <code>ExportImportPortletPreferencesProcessor</code> (in the
+ * <code>com.liferay.exportimport.api</code> module) for the Asset Publisher
+ * portlet. This implementation provides specific export and import capabilities
+ * and routines for processing portlet preferences while exporting or importing
+ * Asset Publisher instances.
+ *
  * @author Mate Thurzo
  */
 @Component(