+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/LiferayOSGiDefaultsPlugin.java
@@ -24,6 +24,7 @@ import com.liferay.gradle.plugins.baseline.BaselineTask;
 import com.liferay.gradle.plugins.cache.CacheExtension;
 import com.liferay.gradle.plugins.cache.CachePlugin;
 import com.liferay.gradle.plugins.cache.task.TaskCache;
+import com.liferay.gradle.plugins.defaults.internal.FindSecurityBugsPlugin;
 import com.liferay.gradle.plugins.defaults.internal.LiferayRelengPlugin;
 import com.liferay.gradle.plugins.defaults.internal.WhipDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.internal.util.BackupFilesBuildAdapter;
@@ -184,6 +185,9 @@ import org.gradle.util.GUtil;
  */
 public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 
+	public static final String ASPECTJ_WEAVER_CONFIGURATION_NAME =
+		"aspectJWeaver";
+
 	public static final String COMMIT_CACHE_TASK_NAME = "commitCache";
 
 	public static final String COPY_LIBS_TASK_NAME = "copyLibs";
@@ -271,6 +275,8 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 
 			WhipDefaultsPlugin.INSTANCE.apply(project);
 
+			Configuration aspectJWeaverConfiguration =
+				_addConfigurationAspectJWeaver(project);
 			Configuration portalConfiguration = GradleUtil.getConfiguration(
 				project, LiferayBasePlugin.PORTAL_CONFIGURATION_NAME);
 			Configuration portalTestConfiguration = _addConfigurationPortalTest(
@@ -284,6 +290,11 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 				project, portalConfiguration, portalTestConfiguration);
 			_configureSourceSetTestIntegration(
 				project, portalConfiguration, portalTestConfiguration);
+			_configureTaskTestAspectJWeaver(
+				project, JavaPlugin.TEST_TASK_NAME, aspectJWeaverConfiguration);
+			_configureTaskTestAspectJWeaver(
+				project, TestIntegrationBasePlugin.TEST_INTEGRATION_TASK_NAME,
+				aspectJWeaverConfiguration);
 		}
 
 		Task baselineTask = GradleUtil.getTask(
@@ -333,7 +344,7 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		configureRepositories(project);
 		_configureSourceSetMain(project);
 		_configureTaskJar(project, testProject);
-		_configureTaskJavadoc(project);
+		_configureTaskJavadoc(project, portalRootDir);
 		_configureTaskTest(project);
 		_configureTaskTestIntegration(project);
 		_configureTaskTlddoc(project, portalRootDir);
@@ -490,6 +501,30 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		}
 	}
 
+	private Configuration _addConfigurationAspectJWeaver(
+		final Project project) {
+
+		Configuration configuration = GradleUtil.addConfiguration(
+			project, ASPECTJ_WEAVER_CONFIGURATION_NAME);
+
+		configuration.defaultDependencies(
+			new Action<DependencySet>() {
+
+				@Override
+				public void execute(DependencySet dependencySet) {
+					_addDependenciesAspectJWeaver(project);
+				}
+
+			});
+
+		configuration.setDescription(
+			"Configures AspectJ Weaver to apply to the test tasks.");
+		configuration.setTransitive(false);
+		configuration.setVisible(false);
+
+		return configuration;
+	}
+
 	private Configuration _addConfigurationPortalTest(Project project) {
 		Configuration configuration = GradleUtil.addConfiguration(
 			project, PORTAL_TEST_CONFIGURATION_NAME);
@@ -500,6 +535,12 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		configuration.setVisible(false);
 
 		return configuration;
+	}
+
+	private void _addDependenciesAspectJWeaver(Project project) {
+		GradleUtil.addDependency(
+			project, ASPECTJ_WEAVER_CONFIGURATION_NAME, "org.aspectj",
+			"aspectjweaver", "1.8.9");
 	}
 
 	private void _addDependenciesPmd(Project project) {
@@ -1015,6 +1056,8 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		if (FileUtil.exists(project, "xsd")) {
 			GradleUtil.applyPlugin(project, XSDBuilderPlugin.class);
 		}
+
+		FindSecurityBugsPlugin.INSTANCE.apply(project);
 	}
 
 	private void _applyVersionOverrideJson(Project project, String fileName)
@@ -1958,7 +2001,7 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 	}
 
 	private void _configureTaskFindBugs(FindBugs findBugs) {
-		findBugs.setMaxHeapSize("1g");
+		findBugs.setMaxHeapSize("3g");
 
 		FindBugsReports findBugsReports = findBugs.getReports();
 
@@ -2064,12 +2107,12 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		compileOptions.setWarnings(false);
 	}
 
-	private void _configureTaskJavadoc(Project project) {
+	private void _configureTaskJavadoc(Project project, File portalRootDir) {
 		Javadoc javadoc = (Javadoc)GradleUtil.getTask(
 			project, JavaPlugin.JAVADOC_TASK_NAME);
 
 		_configureTaskJavadocFilter(javadoc);
-		_configureTaskJavadocOptions(javadoc);
+		_configureTaskJavadocOptions(javadoc, portalRootDir);
 		_configureTaskJavadocTitle(javadoc);
 
 		JavaVersion javaVersion = JavaVersion.current();
@@ -2133,7 +2176,9 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		}
 	}
 
-	private void _configureTaskJavadocOptions(Javadoc javadoc) {
+	private void _configureTaskJavadocOptions(
+		Javadoc javadoc, File portalRootDir) {
+
 		StandardJavadocDocletOptions standardJavadocDocletOptions =
 			(StandardJavadocDocletOptions)javadoc.getOptions();
 
@@ -2164,9 +2209,6 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		}
 
 		standardJavadocDocletOptions.tags("generated");
-
-		File portalRootDir = GradleUtil.getRootDir(
-			project.getRootProject(), "portal-impl");
 
 		if (portalRootDir != null) {
 			File stylesheetFile = new File(
@@ -2345,6 +2387,32 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 		_configureTaskTestLogging(test);
 	}
 
+	private void _configureTaskTestAspectJWeaver(
+		Project project, String taskName,
+		final Configuration aspectJWeaverConfiguration) {
+
+		Test test = (Test)GradleUtil.getTask(project, taskName);
+
+		test.doFirst(
+			new Action<Task>() {
+
+				@Override
+				public void execute(Task task) {
+					Test test = (Test)task;
+
+					test.jvmArgs(
+						"-javaagent:" +
+							FileUtil.getAbsolutePath(
+								aspectJWeaverConfiguration.getSingleFile()));
+				}
+
+			});
+
+		test.systemProperty(
+			"org.aspectj.weaver.loadtime.configuration",
+			"com/liferay/aspectj/modules/aop.xml");
+	}
+
 	private void _configureTaskTestIgnoreFailures(Test test) {
 		test.setIgnoreFailures(true);
 	}
@@ -2435,8 +2503,7 @@ public class LiferayOSGiDefaultsPlugin implements Plugin<Project> {
 			Map<String, Object> args = new HashMap<>();
 
 			args.put("dir", taskCache.getCacheDir());
-			args.put(
-				"includes", Arrays.asList("config.json", "**/*.js"));
+			args.put("includes", Arrays.asList("config.json", "**/*.js"));
 
 			FileTree fileTree = project.fileTree(args);