+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/BaseSourceProcessor.java
@@ -484,7 +484,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		String line, String fileName, String absolutePath, int lineCount,
 		boolean javaSource) {
 
-		if (isExcludedPath(getRunOutsidePortalExcludes(), absolutePath)) {
+		if (isExcludedPath(RUN_OUTSIDE_PORTAL_EXCLUDES, absolutePath)) {
 			return;
 		}
 
@@ -651,7 +651,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		String line, String fileName, String absolutePath, int lineCount) {
 
 		if (!portalSource || fileName.endsWith("ResourceBundleUtil.java") ||
-			isExcludedPath(getRunOutsidePortalExcludes(), absolutePath)) {
+			isExcludedPath(RUN_OUTSIDE_PORTAL_EXCLUDES, absolutePath)) {
 
 			return;
 		}
@@ -792,7 +792,9 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 		String copyright = _copyright;
 
-		if (fileName.endsWith(".vm") || Validator.isNull(copyright)) {
+		if (fileName.endsWith(".tpl") || fileName.endsWith(".vm") ||
+			Validator.isNull(copyright)) {
+
 			return content;
 		}
 
@@ -1400,9 +1402,9 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			String javaClassName, String packagePath, File file,
 			String fileName, String absolutePath, String content,
 			String javaClassContent, int javaClassLineCount, String indent,
-			List<String> checkJavaFieldTypesExcludes,
-			List<String> javaTermSortExcludes,
-			List<String> testAnnotationsExcludes)
+			String checkJavaFieldTypesExcludesProperty,
+			String javaTermSortExcludesProperty,
+			String testAnnotationsExcludesProperty)
 		throws Exception {
 
 		JavaSourceProcessor javaSourceProcessor = null;
@@ -1424,8 +1426,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 		String newJavaClassContent = javaClass.formatJavaTerms(
 			getAnnotationsExclusions(), getImmutableFieldTypes(),
-			checkJavaFieldTypesExcludes, javaTermSortExcludes,
-			testAnnotationsExcludes);
+			checkJavaFieldTypesExcludesProperty, javaTermSortExcludesProperty,
+			testAnnotationsExcludesProperty);
 
 		if (!javaClassContent.equals(newJavaClassContent)) {
 			return StringUtil.replaceFirst(
@@ -2430,19 +2432,6 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			GetterUtil.getString(getProperty(key)), StringPool.COMMA);
 	}
 
-	protected List<String> getRunOutsidePortalExcludes() {
-		if (_runOutsidePortalExcludes != null) {
-			return _runOutsidePortalExcludes;
-		}
-
-		List<String> runOutsidePortalExcludes = getPropertyList(
-			"run.outside.portal.excludes");
-
-		_runOutsidePortalExcludes = runOutsidePortalExcludes;
-
-		return _runOutsidePortalExcludes;
-	}
-
 	protected boolean hasMissingParentheses(String s) {
 		if (Validator.isNull(s)) {
 			return false;
@@ -2595,19 +2584,26 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return matcher.matches();
 	}
 
-	protected boolean isExcludedPath(List<String> excludes, String path) {
-		return isExcludedPath(excludes, path, -1);
+	protected boolean isExcludedPath(String property, String path) {
+		return isExcludedPath(property, path, -1);
 	}
 
 	protected boolean isExcludedPath(
-		List<String> excludes, String path, int lineCount) {
+		String property, String path, int lineCount) {
 
-		return isExcludedPath(excludes, path, lineCount, null);
+		return isExcludedPath(property, path, lineCount, null);
 	}
 
 	protected boolean isExcludedPath(
-		List<String> excludes, String path, int lineCount,
-		String javaTermName) {
+		String property, String path, int lineCount, String javaTermName) {
+
+		List<String> excludes = _exclusionPropertiesMap.get(property);
+
+		if (excludes == null) {
+			excludes = getPropertyList(property);
+
+			_exclusionPropertiesMap.put(property, excludes);
+		}
 
 		if (ListUtil.isEmpty(excludes)) {
 			return false;
@@ -2797,11 +2793,11 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return line;
 	}
 
-	protected String sortMethodCalls(
+	protected String sortMethodCall(
 		String content, String methodName, String... variableTypeRegexStrings) {
 
 		Pattern codeBlockPattern = Pattern.compile(
-			"(\t*(\\w*)\\." + methodName + "\\(\\s*\".*?\\);\n)+",
+			"(\t+(\\w*)\\." + methodName + "\\(\\s*\".*?\\);\n)+",
 			Pattern.DOTALL);
 
 		Matcher codeBlockMatcher = codeBlockPattern.matcher(content);
@@ -2852,6 +2848,22 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 				previousPutOrSetParameterName = putOrSetParameterName;
 			}
 		}
+
+		return content;
+	}
+
+	protected String sortMethodCalls(String absolutePath, String content) {
+		if (isExcludedPath(METHOD_CALL_SORT_EXCLUDES, absolutePath)) {
+			return content;
+		}
+
+		content = sortMethodCall(
+			content, "add", "ConcurrentSkipListSet<.*>", "HashSet<.*>",
+			"TreeSet<.*>");
+		content = sortMethodCall(
+			content, "put", "ConcurrentHashMap<.*>", "HashMap<.*>",
+			"JSONObject", "TreeMap<.*>");
+		content = sortMethodCall(content, "setAttribute");
 
 		return content;
 	}
@@ -2984,6 +2996,12 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return line;
 	}
 
+	protected static final String METHOD_CALL_SORT_EXCLUDES =
+		"method.call.sort.excludes";
+
+	protected static final String RUN_OUTSIDE_PORTAL_EXCLUDES =
+		"run.outside.portal.excludes";
+
 	protected static Pattern applyLangMergerPluginPattern = Pattern.compile(
 		"^apply[ \t]+plugin[ \t]*:[ \t]+\"com.liferay.lang.merger\"$",
 		Pattern.MULTILINE);
@@ -3085,9 +3103,29 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		if (getFile("portal-impl", PORTAL_MAX_DIR_LEVEL) != null) {
 			return true;
 		}
-		else {
-			return false;
+
+		// Subrepositories should be treated as portal
+
+		String baseDirAbsolutePath = getAbsolutePath(
+			sourceFormatterArgs.getBaseDirName());
+
+		int x = baseDirAbsolutePath.length();
+
+		for (int i = 0; i < 2; i++) {
+			x = baseDirAbsolutePath.lastIndexOf(CharPool.FORWARD_SLASH, x - 1);
+
+			if (x == -1) {
+				return false;
+			}
+
+			String dirName = baseDirAbsolutePath.substring(x + 1);
+
+			if (dirName.startsWith("com-liferay-")) {
+				return true;
+			}
 		}
+
+		return false;
 	}
 
 	private String _normalizePattern(String originalPattern) {
@@ -3119,6 +3157,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 	private final Pattern _emptyLineInNestedTagsPattern2 = Pattern.compile(
 		"\n(\t*)(.*>)\n\n(\t*)</.*(\n|$)");
 	private String[] _excludes;
+	private Map<String, List<String>> _exclusionPropertiesMap = new HashMap<>();
 	private SourceMismatchException _firstSourceMismatchException;
 	private Set<String> _immutableFieldTypes;
 	private ComparableVersion _mainReleaseComparableVersion;
@@ -3132,7 +3171,6 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 	private List<String> _pluginsInsideModulesDirectoryNames;
 	private Properties _portalLanguageProperties;
 	private Properties _properties;
-	private List<String> _runOutsidePortalExcludes;
 	private SourceFormatterHelper _sourceFormatterHelper;
 	private Map<String, Set<SourceFormatterMessage>>
 		_sourceFormatterMessagesMap = new ConcurrentHashMap<>();
@@ -3166,6 +3204,17 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 			if (matcher.find()) {
 				putOrSetParameterName2 = matcher.replaceAll(StringPool.BLANK);
+			}
+
+			if (putOrSetParameterName1.matches("\".*\"") &&
+				putOrSetParameterName2.matches("\".*\"")) {
+
+				String strippedQuotes1 = putOrSetParameterName1.substring(
+					1, putOrSetParameterName1.length() - 1);
+				String strippedQuotes2 = putOrSetParameterName2.substring(
+					1, putOrSetParameterName2.length() - 1);
+
+				return super.compare(strippedQuotes1, strippedQuotes2);
 			}
 
 			int value = super.compare(